gateway_service/requirements.txt
fastapi
httpx
PyJWT
pydantic
uvicorn

# Service Registry
python-consul==1.1.0

# Distributed Tracing
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0
opentelemetry-exporter-jaeger>=1.15.0
opentelemetry-instrumentation-flask>=0.37b0
opentelemetry-instrumentation-requests>=0.37b0
opentelemetry-instrumentation-sqlalchemy>=0.37b0

# Monitoring
prometheus-flask-exporter>=0.22.4

# API Versioning
apispec>=6.0.0
flask-apispec>=0.11.0
